Create LICENSE
Would is be possible to add a LICENSE file? According to the package.json the license is MIT, so I wrote up this one based on the date of the first commit. I'd love, love, love to use this theme instead of Dracula but we have a strict policy requiring license files in the codebases of VS Code extensions.